@@ -3,23 +3,23 @@ apply plugin: 'com.android.application'
 apply from: project.property("cahue.signing") + "/cahue.gradle";
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.2'
+    compileSdkVersion 27
+    buildToolsVersion "27.0.0"
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 25
+        minSdkVersion 21
+        targetSdkVersion 27
         applicationId "com.whereismycar"
         versionCode 83
         versionName "3.7.0"
         signingConfig signingConfigs.wimc
-        multiDexEnabled true
 
         buildConfigField "String", "BACKEND_URL", "\"glossy-radio.appspot.com\""
 
     }
 
-    dexOptions {
-        javaMaxHeapSize "4g"
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     buildTypes {
@@ -42,30 +42,30 @@ dependencies {
         exclude group: "com.google.android.gms"
     }
 
-    compile 'com.android.support:appcompat-v7:25.2.0'
-    compile 'com.android.support:support-v13:25.2.0'
-    compile 'com.android.support:cardview-v7:25.2.0'
-    compile 'com.android.support:recyclerview-v7:25.2.0'
-    compile 'com.android.support:design:25.2.0'
-    compile 'com.google.android.gms:play-services-auth:10.2.0'
-    compile 'com.google.android.gms:play-services-plus:10.2.0'
-    compile 'com.google.android.gms:play-services-maps:10.2.0'
-    compile 'com.google.android.gms:play-services-location:10.2.0'
-    compile 'com.google.android.gms:play-services-wallet:10.2.0'
-    compile 'com.google.android.gms:play-services-gcm:10.2.0'
-    compile 'com.google.android.gms:play-services-analytics:10.2.0'
-    compile 'com.google.android.gms:play-services-ads:10.2.0'
-    compile 'de.hdodenhof:circleimageview:1.2.2'
+    compile 'com.android.support:appcompat-v7:27.0.0'
+    compile 'com.android.support:support-v13:27.0.0'
+    compile 'com.android.support:cardview-v7:27.0.0'
+    compile 'com.android.support:recyclerview-v7:27.0.0'
+    compile 'com.android.support:design:27.0.0'
+    compile 'com.google.android.gms:play-services-awareness:11.6.0'
+    compile 'com.google.android.gms:play-services-auth:11.6.0'
+    compile 'com.google.android.gms:play-services-plus:11.6.0'
+    compile 'com.google.android.gms:play-services-maps:11.6.0'
+    compile 'com.google.android.gms:play-services-location:11.6.0'
+    compile 'com.google.android.gms:play-services-wallet:11.6.0'
+    compile 'com.google.android.gms:play-services-gcm:11.6.0'
+    compile 'com.google.android.gms:play-services-analytics:11.6.0'
+    compile 'com.google.android.gms:play-services-ads:11.6.0'
+    compile 'de.hdodenhof:circleimageview:1.3.0'
     compile 'com.mcxiaoke.volley:library:1.0.19@aar'
     compile 'pl.droidsonroids.gif:android-gif-drawable:1.2.3'
 
-    compile ('com.facebook.android:facebook-android-sdk:4.+'){
-        exclude group: "com.google.android.gms"
-        exclude group: "com.android.support"
+    compile ('com.facebook.android:facebook-android-sdk:4.28.0'){
+        exclude group: 'com.android.support'
     }
-    compile('com.facebook.android:audience-network-sdk:4.+') {
-        exclude group: "com.google.android.gms"
-        exclude group: "com.android.support"
+    compile ('com.facebook.android:audience-network-sdk:4.26.1'){
+        exclude group: 'com.android.support'
     }
+
 
 }
