@@ -2,9 +2,7 @@ package com.cahue.iweco;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.v4.view.ViewCompat;
 import android.support.v7.widget.Toolbar;
-import android.view.View;
 
 /**
  * Created by f.gatti.gomez on 08/07/15.
@@ -17,31 +15,12 @@ public class SettingsActivity extends Activity {
 
         setContentView(R.layout.activity_preferences);
 
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar_actionbar);
+        Toolbar toolbar =  findViewById(R.id.toolbar_actionbar);
         toolbar.setTitle(R.string.preferences);
         toolbar.setNavigationIcon(R.drawable.ic_action_navigation_arrow_back);
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
-        ViewCompat.setElevation(toolbar, getResources().getDimension(R.dimen.elevation));
-
-//        // Display the fragment as the main content.
-//        getFragmentManager().beginTransaction()
-//                .replace(android.R.id.content, new PreferencesFragment())
-//                .commit();
+        toolbar.setNavigationOnClickListener(v -> finish());
     }
 
-//    /**
-//     * Populate the activity with the top-level headers.
-//     */
-//    @Override
-//    public void onBuildHeaders(List<Header> target) {
-//        loadHeadersFromResource(R.xml.preference_headers, target);
-//    }
-//
 
 }
 
