@@ -1,12 +1,10 @@
 package com.cahue.iweco.activityrecognition;
 
-import android.content.Context;
 import android.location.Location;
 import android.os.Bundle;
+import android.support.annotation.IntDef;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.view.LayoutInflater;
-import android.view.View;
 
 import com.cahue.iweco.AbstractMarkerDelegate;
 import com.cahue.iweco.DetailsFragment;
@@ -16,6 +14,7 @@ import com.cahue.iweco.cars.CarsSync;
 import com.cahue.iweco.cars.database.CarDatabase;
 import com.cahue.iweco.model.Car;
 import com.cahue.iweco.model.ParkingSpot;
+import com.cahue.iweco.util.Tracking;
 import com.google.android.gms.maps.CameraUpdate;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
@@ -25,6 +24,10 @@ import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
 import com.google.maps.android.ui.IconGenerator;
 
+import java.lang.annotation.Retention;
+
+import static java.lang.annotation.RetentionPolicy.SOURCE;
+
 /**
  * Delegate to show a marker where a car might be parked, based on activity recognition
  */
@@ -32,14 +35,17 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
 
     public static final String FRAGMENT_TAG = "POSSIBLE_PARKED_CAR_DELEGATE";
 
-    private static final String ARG_SPOT = "spot";
+    public static final int RECENT = 0;
+    public static final int NOT_SO_RECENT = 1;
 
+    private static final String ARG_SPOT = "spot";
+    private static final String ARG_RECENCY = "recency";
+    @Recency
+    private int recency;
     @Nullable
     private ParkingSpot spot;
-
     @Nullable
     private Marker marker;
-
     private IconGenerator iconGenerator;
     private boolean following;
     private CarDatabase database;
@@ -50,10 +56,11 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
     }
 
     @NonNull
-    public static PossibleParkedCarDelegate newInstance(ParkingSpot spot) {
+    public static PossibleParkedCarDelegate newInstance(ParkingSpot spot, @Recency int recency) {
         PossibleParkedCarDelegate fragment = new PossibleParkedCarDelegate();
         Bundle args = new Bundle();
         args.putParcelable(ARG_SPOT, spot);
+        args.putInt(ARG_RECENCY, recency);
         fragment.setArguments(args);
         return fragment;
     }
@@ -63,11 +70,21 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
         super.onCreate(savedInstanceState);
         database = CarDatabase.getInstance();
         this.spot = getArguments().getParcelable(ARG_SPOT);
+        //noinspection WrongConstant
+        this.recency = getArguments().getInt(ARG_RECENCY);
+
         iconGenerator = new IconGenerator(getActivity());
-        LayoutInflater myInflater = (LayoutInflater) getActivity().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        View contentView = myInflater.inflate(R.layout.marker_possible_view, null, false);
-        iconGenerator.setBackground(null);
-        iconGenerator.setContentView(contentView);
+        iconGenerator.setContentRotation(-90);
+        switch (recency) {
+            case RECENT:
+                iconGenerator.setColor(getActivity().getResources().getColor(R.color.lightest_gray));
+                iconGenerator.setTextAppearance(getActivity(), R.style.Marker_PossibleCar);
+                break;
+            case NOT_SO_RECENT:
+                iconGenerator.setColor(getActivity().getResources().getColor(R.color.lightest_gray));
+                iconGenerator.setTextAppearance(getActivity(), R.style.Marker_PossibleCar10);
+                break;
+        }
     }
 
     @Override
@@ -84,7 +101,8 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
 
         marker = getMap().addMarker(new MarkerOptions()
                 .position(spot.getLatLng())
-                .icon(BitmapDescriptorFactory.fromBitmap(iconGenerator.makeIcon())));
+                .icon(BitmapDescriptorFactory.fromBitmap(iconGenerator.makeIcon("Possible parked car")))
+                .anchor(iconGenerator.getAnchorU(), iconGenerator.getAnchorV()));
 
     }
 
@@ -102,6 +120,8 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
     public boolean onMarkerClick(@NonNull Marker marker) {
         if (marker.equals(this.marker)) {
             activate();
+
+            Tracking.sendEvent(Tracking.CATEGORY_MAP, Tracking.ACTION_POSSIBLE_CAR_SELECTED, Tracking.LABEL_SELECTED_FROM_MARKER);
             return true;
         }
         return false;
@@ -119,7 +139,6 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
         DetailsFragment detailsFragment = detailsViewManager.getDetailsFragment();
         return detailsFragment != null && detailsFragment instanceof PossibleSetCarDetailsFragment;
     }
-
 
     private void clearMarker() {
         if (marker != null) {
@@ -160,5 +179,10 @@ public class PossibleParkedCarDelegate extends AbstractMarkerDelegate implements
         if (marker != null)
             marker.remove();
         detailsViewManager.hideDetails();
+    }
+
+    @Retention(SOURCE)
+    @IntDef({RECENT, NOT_SO_RECENT})
+    public @interface Recency {
     }
 }
