@@ -8,6 +8,7 @@ import android.database.sqlite.SQLiteDatabase;
 import android.location.Location;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 
 import com.cahue.iweco.Constants;
@@ -258,7 +259,7 @@ public class CarDatabase {
 
         Intent intent = new Intent(Constants.INTENT_CAR_UPDATED);
         intent.putExtra(Constants.EXTRA_CAR_ID, car.id);
-        context.sendBroadcast(intent);
+        LocalBroadcastManager.getInstance(context).sendBroadcast(intent);
     }
 
     /**
@@ -576,23 +577,38 @@ public class CarDatabase {
         try {
             ContentValues values = createSpotContentValues(spot);
 
-            database.insertWithOnConflict(TABLE_POSSIBLE_SPOTS, COLUMN_TIME, values, SQLiteDatabase.CONFLICT_REPLACE);
-
             Cursor cursor = database.query(TABLE_POSSIBLE_SPOTS,
-                    new String[]{COLUMN_TIME},
+                    SPOT_PROJECTION,
                     null,
                     null, null, null,
                     COLUMN_TIME + " DESC");
 
-            List<Long> parkingTimes = new ArrayList<>();
+            List<ParkingSpot> previousSpots = new ArrayList<>();
             while (cursor.moveToNext()) {
-                parkingTimes.add(cursor.getLong(0));
+                previousSpots.add(cursorToSpot(cursor));
             }
             cursor.close();
 
-            if (parkingTimes.size() > MAX_POSSIBLE_SPOTS) {
-                for (Long time : parkingTimes.subList(MAX_POSSIBLE_SPOTS, parkingTimes.size())) {
-                    database.delete(TABLE_POSSIBLE_SPOTS, COLUMN_TIME + " = '" + time.toString() + "'", null);
+            database.insertWithOnConflict(TABLE_POSSIBLE_SPOTS, COLUMN_TIME, values, SQLiteDatabase.CONFLICT_REPLACE);
+
+            // remove spots that are too close
+            for (ParkingSpot prevSpot : previousSpots) {
+                float distances[] = new float[3];
+                Location.distanceBetween(
+                        spot.getLatLng().latitude,
+                        spot.getLatLng().longitude,
+                        prevSpot.getLatLng().latitude,
+                        prevSpot.getLatLng().longitude,
+                        distances);
+
+                if (distances[0] < 25)
+                    database.delete(TABLE_POSSIBLE_SPOTS, COLUMN_TIME + " = '" + prevSpot.getTime().getTime() + "'", null);
+            }
+
+            // remove stale spots
+            if (previousSpots.size() > MAX_POSSIBLE_SPOTS - 1) {
+                for (ParkingSpot prevSpot : previousSpots.subList(MAX_POSSIBLE_SPOTS - 1, previousSpots.size())) {
+                    database.delete(TABLE_POSSIBLE_SPOTS, COLUMN_TIME + " = '" + prevSpot.getTime().getTime() + "'", null);
                 }
             }
 
