@@ -11,6 +11,7 @@ import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.util.Log;
@@ -135,14 +136,14 @@ public class LoginActivity extends AppCompatActivity implements LoginAsyncTask.L
 
         mFacebookLoginButton = (Button) findViewById(R.id.facebook_login_button);
 
-        // Facebook callback registration
+        // Facebook callbacks registration
         mFacebookCallbackManager = CallbackManager.Factory.create();
 
         final LoginManager loginManager = LoginManager.getInstance();
         mFacebookLoginButton.setOnClickListener(new OnClickListener() {
             @Override
             public void onClick(View v) {
-                loginManager.logInWithReadPermissions(LoginActivity.this, Arrays.asList("user_friends", "email"));
+                loginManager.logInWithReadPermissions(LoginActivity.this, Arrays.asList("public_profile", "user_friends", "email"));
                 setLoading(true);
             }
         });
@@ -490,7 +491,7 @@ public class LoginActivity extends AppCompatActivity implements LoginAsyncTask.L
                                     object.getString("email"),
                                     String.format("https://graph.facebook.com/%s/picture", object.getString("id")));
                             Intent intent = new Intent(Constants.INTENT_USER_INFO_UPDATE);
-                            sendBroadcast(intent);
+                            LocalBroadcastManager.getInstance(LoginActivity.this).sendBroadcast(intent);
                         } catch (JSONException e) {
                             e.printStackTrace();
                         }
